refactor(tests): refactorizar tests de processors - dividir archivos >300 líneas
- Dividir payment.processor.spec.ts (787 líneas) en 3 archivos:
  * payment.processor.core.spec.ts (291 líneas)
  * payment.processor.edge-cases.spec.ts (198 líneas)
  * payment.processor.validations-tracking.spec.ts (201 líneas)

- Dividir inventory.processor.spec.ts (781 líneas) en 3 archivos:
  * inventory.processor.core.spec.ts (323 líneas)
  * inventory.processor.edge-cases.spec.ts (202 líneas)
  * inventory.processor.tracking-metrics.spec.ts (187 líneas)

- Dividir notification.processor.spec.ts (762 líneas) en 3 archivos:
  * notification.processor.core.spec.ts (266 líneas)
  * notification.processor.error-tracking.spec.ts (224 líneas)
  * notification.processor.features.spec.ts (175 líneas)

- Dividir order-processing.processor.spec.ts (652 líneas) en 3 archivos:
  * order-processing.processor.core.spec.ts (176 líneas)
  * order-processing.processor.error-handling.spec.ts (131 líneas)
  * order-processing.processor.edge-cases-metrics.spec.ts (207 líneas)

- Crear helpers centralizados:
  * helpers/payment-processor.test-helpers.ts
  * helpers/inventory-processor.test-helpers.ts
  * helpers/notification-processor.test-helpers.ts
  * helpers/order-processing-processor.test-helpers.ts

Tests: 162 passing (100%), Promedio: 216 líneas/archivo
